Revert "move AuxRandElementsGenerator to air crate"
This reverts commit 367a5d2f161519dbc243b55f27f028155627c1b8.
