+++ air/src/air/mod.rs
@@ -8,9 +8,6 @@ use alloc::{collections::BTreeMap, vec::Vec};
 use crypto::{RandomCoin, RandomCoinError};
 use math::{fft, ExtensibleField, ExtensionOf, FieldElement, StarkField, ToElements};
 
-mod aux;
-pub use aux::{AuxRandElementsGenerator, DefaultAuxRandElementsGenerator};
-
 mod trace_info;
 pub use trace_info::TraceInfo;
