+++ air/src/lib.rs
@@ -42,9 +42,8 @@ pub use options::{FieldExtension, ProofOptions};
 
 mod air;
 pub use air::{
-    Air, AirContext, Assertion, AuxRandElementsGenerator, BoundaryConstraint,
-    BoundaryConstraintGroup, BoundaryConstraints, ConstraintCompositionCoefficients,
-    ConstraintDivisor, DeepCompositionCoefficients, DefaultAuxRandElementsGenerator,
+    Air, AirContext, Assertion, BoundaryConstraint, BoundaryConstraintGroup, BoundaryConstraints,
+    ConstraintCompositionCoefficients, ConstraintDivisor, DeepCompositionCoefficients,
     EvaluationFrame, LagrangeConstraintsCompositionCoefficients, LagrangeKernelBoundaryConstraint,
     LagrangeKernelConstraints, LagrangeKernelEvaluationFrame, LagrangeKernelTransitionConstraints,
     TraceInfo, TransitionConstraintDegree, TransitionConstraints,
