+++ examples/src/rescue_raps/mod.rs
@@ -164,7 +164,7 @@ where
         };
         let acceptable_options =
             winterfell::AcceptableOptions::OptionSet(vec![proof.options().clone()]);
-        let aux_rand_elements_generator = DefaultAuxRandElementsGenerator::new(3);
+        let aux_verifier = DefaultAuxRandElementsGenerator::new(3);
 
         match proof.options().field_extension() {
             FieldExtension::None => {
@@ -174,9 +174,7 @@ where
                     _,
                     H,
                     DefaultRandomCoin<H>,
-                >(
-                    proof, aux_rand_elements_generator, pub_inputs, &acceptable_options
-                )
+                >(proof, aux_verifier, pub_inputs, &acceptable_options)
             }
 
             FieldExtension::Quadratic => {
@@ -186,9 +184,7 @@ where
                     _,
                     H,
                     DefaultRandomCoin<H>,
-                >(
-                    proof, aux_rand_elements_generator, pub_inputs, &acceptable_options
-                )
+                >(proof, aux_verifier, pub_inputs, &acceptable_options)
             }
             FieldExtension::Cubic => {
                 winterfell::verify_with_aux_trace::<
@@ -197,9 +193,7 @@ where
                     _,
                     H,
                     DefaultRandomCoin<H>,
-                >(
-                    proof, aux_rand_elements_generator, pub_inputs, &acceptable_options
-                )
+                >(proof, aux_verifier, pub_inputs, &acceptable_options)
             }
         }
     }
