+++ verifier/src/aux_verifier.rs
