+++ winterfell/src/lib.rs
@@ -588,8 +588,6 @@
 #[cfg(test)]
 extern crate std;
 
-pub use air::{AuxRandElementsGenerator, DefaultAuxRandElementsGenerator};
-
 pub use prover::{
     crypto, iterators, math, matrix, Air, AirContext, Assertion, AuxParams, AuxProof,
     AuxRandElements, AuxTraceBuilder, AuxTraceWithMetadata, BoundaryConstraint,
@@ -600,7 +598,10 @@ pub use prover::{
     Serializable, SliceReader, StarkDomain, StarkProof, Trace, TraceInfo, TraceLde, TracePolyTable,
     TraceTable, TraceTableFragment, TransitionConstraintDegree,
 };
-pub use verifier::{verify, verify_with_aux_trace, AcceptableOptions, VerifierError};
+pub use verifier::{
+    verify, verify_with_aux_trace, AcceptableOptions, DefaultAuxRandElementsGenerator,
+    VerifierError,
+};
 
 #[cfg(test)]
 mod tests;
